+++ tests/render_spec.rs
@@ -58,6 +58,8 @@ fn should_render_attribute(#[case] attr: Attribute, #[case] expected: &str) {
 #[case(raw("<my-component></my-component>"), "<my-component></my-component>")]
 #[case(raw_unsafe("<my-component></my-component>".to_string()), "<my-component></my-component>")]
 #[case(meta([("foo", "bar").into()]), "<meta foo=\"bar\">")]
+#[case(link([("foo", "bar").into()]), "<link foo=\"bar\">")]
+#[case(script([("foo", "bar").into()], ["alert('hello');".into()]), "<script foo=\"bar\">alert('hello');</script>")]
 #[case(
     meta_viewport(),
     "<meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">"